More cleaning & file reorganisation
