+++ setup_requirements.py
@@ -44,7 +44,6 @@ install_requires = [
     'SQLAlchemy==1.0.12',  # upgrade to SQLalchemy 1.1.5 does break tests, see #465
     'SQLAlchemy-Utils==0.31.2',
     'alembic>=0.9.2',
-    'alembic-verify>=0.1.4',
     'ujson==1.35',
     'enum34==1.1.6',
     'voluptuous==0.8.11',
@@ -121,7 +120,8 @@ extras_require = {
     # Requirements for testing
     'testing': [
         'mock',
-        'pgtest'
+        'pgtest',
+        'sqlalchemy-diff>=0.1.3'
     ],
     'dev_precommit': [
         'pre-commit',
